Massive CGate & Core refactoring

+++ cgate/src/main/scala/com/ergodicity/cgate/SysEvent.scala
@@ -3,7 +3,7 @@ package com.ergodicity.cgate
 sealed trait SysEvent {
   def eventId: Long
 
-  def sessionId: Long
+  def sessionId: Int
 }
 
 object SysEvent {
